// src/redux/tag-types.ts
/* eslint-disable no-unused-vars */
// eslint-disable-next-line no-shadow
export enum tagTypes {
  user = 'user',
  services = 'services',
  admin = 'admin',
  chatbot = 'chatbot',
  auth = 'auth',
  knowledgeBase = 'knowledgeBase',
  sampleQustion = 'sampleQustion',
  userChatbot = 'userChatbot',
  widgetConfig = 'widgetConfig',
  feedback = 'feedback',
  prompt = 'prompt',
  analytics = 'analytics',
}

export const tagTypesList = [
  tagTypes.user,
  tagTypes.auth,
  tagTypes.services,
  tagTypes.admin,
  tagTypes.chatbot,
  tagTypes.knowledgeBase,
  tagTypes.sampleQustion,
  tagTypes.userChatbot,
  tagTypes.widgetConfig,
  tagTypes.feedback,
  tagTypes.prompt,
  tagTypes.analytics,
];